# zoo/src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Animal Management API
  description: API to manage animals, clients, orders, and operations.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local Server
paths:
...
  /animals:
    post:
      summary: Add a new animal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_animal_template, sex, origin, price, rent, status, color]
              properties:
                id_animal_template:
                  type: integer
                sex:
                  type: string
                  enum: [male, female]
                origin:
                  type: string
                price:
                  type: number
                rent:
                  type: number
                status:
                  type: string
                  enum: [available, unavailable]
                color:
                  type: string
      responses:
        201:
          description: Animal added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalError'

    get:
      summary: Get all animals
      responses:
        200:
          description: List of animals returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Animal'
  /animals/{id}:
    get:
      summary: Get an animal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Animal details returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Animal'
        404:
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update an animal status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [available, unavailable]
      responses:
        200:
          description: Animal updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete an animal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        204:
          description: Animal deleted successfully
        404:
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    Animal:
      type: object
      properties:
        id:
          type: integer
          example: 1
        id_animal_template:
          type: integer
          example: 1
        sex:
          type: string
          enum: [male, female]
          example: male
        origin:
          type: string
          example: Zoo Paris
        price:
          type: number
          example: 10000.00
        rent:
          type: number
          example: 500.00
        status:
          type: string
          enum: [available, unavailable]
          example: available
        color:
          type: string
          example: golden
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
        data:
          type: object
          nullable: true
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: ERR_001
        message:
          type: string
          example: Invalid request parameters
        details:
          type: array
          items:
            type: string
          nullable: true
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'